Create an ALASCA model

ALASCA initializes an ALASCA model and returns an ALASCA object.

ALASCA(df, formula, effects = NULL, ...)

Arguments

df

Data frame to be analyzed

formula

Regression formula

effects

Vector with effect terms. If NULL, ALASCA will guess (default)

scale_function

Either a custom function or string to define scaling function: sdall, sdref, sdt1, sdreft1

separate_effects

Boolean. Separate effects?

equal_baseline

Set to TRUE to remove interaction between effects

validate

Boolean. If TRUE, give estimates for robustness

n_validation_runs

number of validation runs

validation_method

Choose between bootstrap (default) and "jack-knife"

stratification_column

The column to stratify participants by

save

Save models and plots automatically (default: FALSE)

filename

File name to save model and plots (when save = TRUE)

use_Rfast

Boolean. Defaults to TRUE

p_adjust_method

Method for correcting p values for multiple testing, see p.adjust.methods

participant_column

String. Name of the column containing participant identification

n_validation_folds

Partitions when validating

Value

An ALASCA object

Details

This function builds your ALASCA model. It needs, at least, a data frame and a formula. The effect matrices can be specified with effects, e.g., c("time", "time+group+time:group", "group+time:group").