R6 Class that helps visualizing an ALASCA model

The class contains plot options and the functions for plotting themselves. It can be accessed as obj$splot where obj is an ALASCA object.

Note: The object is typically accessed by plot(obj, ...) where the arguments are passed to the class. The arguments can also be set when the ALASCA model is created by providing arguments with prefix plot. (e.g., plot.my_theme = ...)

Public fields

model

ALASCA model

my_theme

Theme for ggplot2 plots

variable_label

Text label for the y axis

variable

Selected variables to plot

x_label

Text label for the x axis

group_label

Legend title

n_bins

Number of bins for histograms

sort_loadings

String: loading, alpha, vector

ribbon

Boolean. Plot ribbons for uncertainties

bw

Boolean. Plot in gray scale

bw

Boolean. Same as grayscale

bw

Boolean. Same as grayscale

dodgewidth

Validated figures have dodged points to avoid overlap

dpi

Resolution when saving figures. See ggplot2::ggsave()

units

Units for figure sizes. See ggplot2::ggsave()

filetype

File type for saved plots. See ggplot2::ggsave()

loading_group_column

Column for variable groups

loading_group_label

Legend table for variable groups

sort_by_loading_group

Boolean. Sort loadings within loading group

palette_end

Improve contrast by not using the full color space. See scales::viridis_pal()

effect_i

Effect(s) to plot

component

Component(s) to plot

facet_ncol

Number of facet columns. See ggplot2::facet_wrap()

facet_nrow

Number of facet rows. See ggplot2::facet_wrap()

save

Boolean. Whether to save plots

flip_axis

Boolean. If TRUE, variabels are plotted along the y axis

x_angle

Angle for the x labels

labels

Figure labels, see ggpubr::ggarrange()

type

Plot type

Methods

Public methods

AlascaPlot$plot_prediction()

AlascaPlot$plot_effect()

AlascaPlot$plot_effect_score()

AlascaPlot$plot_effect_loading()

AlascaPlot$plot_effect_validation()

AlascaPlot$plot_effect_validation_score()

AlascaPlot$plot_effect_validation_loading()

AlascaPlot$plot_2D_advanced()

AlascaPlot$plot_2D()

AlascaPlot$plot_2D_score()

AlascaPlot$plot_2D_loading_1()

AlascaPlot$plot_2D_loading_2()

AlascaPlot$plot_histogram_score()

AlascaPlot$plot_histogram_loading()

AlascaPlot$plot_histogram()

AlascaPlot$plot_participants()

AlascaPlot$new()

AlascaPlot$call_plot()

AlascaPlot$plot_scree()

AlascaPlot$plot_residuals()

AlascaPlot$plot_covars()

AlascaPlot$capitalize()

AlascaPlot$prettify_covar()

AlascaPlot$get_plot_linetypes()

AlascaPlot$get_plot_shapes()

AlascaPlot$get_plot_palette()

AlascaPlot$get_explained_label()

AlascaPlot$get_levels()

AlascaPlot$get_ref()

AlascaPlot$xflip()

AlascaPlot$clone()

Method plot_prediction()

Usage

AlascaPlot$plot_prediction()

Method plot_effect()

Usage

AlascaPlot$plot_effect()

Method plot_effect_score()

Usage

AlascaPlot$plot_effect_score(effect_i = 1, component = 1)

Method plot_effect_loading()

Usage

AlascaPlot$plot_effect_loading(effect_i = 1, component = 1)

Method plot_effect_validation()

Usage

AlascaPlot$plot_effect_validation()

Method plot_effect_validation_score()

Usage

AlascaPlot$plot_effect_validation_score(effect_i = 1, component = 1)

Method plot_effect_validation_loading()

Usage

AlascaPlot$plot_effect_validation_loading(effect_i = 1, component = 1)

Method plot_2D_advanced()

Usage

AlascaPlot$plot_2D_advanced()

Method plot_2D()

Usage

AlascaPlot$plot_2D()

Method plot_2D_score()

Usage

AlascaPlot$plot_2D_score()

Method plot_2D_loading_1()

Usage

AlascaPlot$plot_2D_loading_1()

Method plot_2D_loading_2()

Usage

AlascaPlot$plot_2D_loading_2()

Method plot_histogram_score()

Usage

AlascaPlot$plot_histogram_score()

Method plot_histogram_loading()

Usage

AlascaPlot$plot_histogram_loading()

Method plot_histogram()

Usage

AlascaPlot$plot_histogram()

Method plot_participants()

Usage

AlascaPlot$plot_participants(effect_i = 1, component = 1)

Method new()

Usage

AlascaPlot$new(model)

Method call_plot()

Usage

AlascaPlot$call_plot(...)

Method plot_scree()

Usage

AlascaPlot$plot_scree()

Method plot_residuals()

Usage

AlascaPlot$plot_residuals()

Method plot_covars()

Usage

AlascaPlot$plot_covars()

Method capitalize()

Usage

AlascaPlot$capitalize(txt)

Method prettify_covar()

Usage

AlascaPlot$prettify_covar(effect, txt)

Method get_plot_linetypes()

Usage

AlascaPlot$get_plot_linetypes()

Method get_plot_shapes()

Usage

AlascaPlot$get_plot_shapes()

Method get_plot_palette()

Usage

AlascaPlot$get_plot_palette()

Method get_explained_label()

Usage

AlascaPlot$get_explained_label(effect_i = 1, component = 1, type = "Score")

Method get_levels()

Usage

AlascaPlot$get_levels(x)

Method get_ref()

Usage

AlascaPlot$get_ref(x)

Method xflip()

Usage

AlascaPlot$xflip(flip = TRUE)

Method clone()

The objects of this class are cloneable with this method.

Usage

AlascaPlot$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.