Get an ALASCA object

ALASCA initializes an ALASCA model and returns an ALASCA object

ALASCA(
  df,
  formula,
  separateTimeAndGroup = TRUE,
  pAdjustMethod = NA,
  participantColumn = "ID",
  validate = FALSE,
  scaleFun = TRUE,
  forceEqualBaseline = FALSE,
  useSumCoding = FALSE,
  method = NA,
  stratificationVector = NA,
  minimizeObject = FALSE,
  plot.xlabel = "Time",
  doDebug = FALSE,
  nValFold = 7,
  nValRuns = 50,
  validateRegression = FALSE,
  validationMethod = "loo",
  validationObject = NA,
  validationParticipants = NA
)

Arguments

df
Data frame to be analyzed

formula
Regression model

separateTimeAndGroup
Logical: should time and group effect be separated?

pAdjustMethod
Method for correcting p values for multiple testing, see p.adjust.methods

participantColumn
String. Name of the column containing participant identification

validate
Logical. If TRUE, give estimates for robustness

scaleFun
If TRUE (default), each variable is scaled to unit SD and zero mean. If FALSE, no scaling is performed. You can also provide a custom scaling function that has the data frame df as input and output

forceEqualBaseline
Set to TRUE (default) to remove interaction between group and first time point

useSumCoding
Set to TRUE to use sum coding instead of contrast coding for group (defaults to FALSE)

method
Defaults to NA where method is either LM or LMM or Rfast, depending on whether your formula contains a random effect or not

stratificationVector
Vector of same length as df that specifies stratification groups during validation. Defaults to NA, where the group column is used.

minimizeObject
Logical. If TRUE, remove unnecessary clutter, optimize for validation

plot.xlabel
Defaults to "Time"

doDebug
Print what happens (default: FALSE)

nValFold
Partitions when validating

nValRuns
number of validation runs

validateRegression
Whether to validate regression predictions or not (only if validate is TRUE)

validationMethod
among loo (leave-one-out, default)

validationObject
Don't worry about me:)

validationParticipants
Don't worry about me:)

Value

An ALASCA object

Details

This function builds your ALASCA model. It needs a data frame containing at least a column identifying participants, a column called time contining time information, a column group containing group information, a column variable containing variable names, and a value column. In addition you need to specify the model you want, and whether you want to separate group and time effects (defaults to TRUE).

Examples

load("PE.Rdata")
model <- ALASCA(df = df, formula = value~time*group + (1|ID))

#> Will use linear mixed models!
#> Using group for stratification.
#> Scaling data...
#> Calculating  LMM  coefficients...
#> Finished calculating regression coefficients!
#> Calculating predictions from regression models...
#> Finished calculating predictions from regression models!
#> Calculating effect matrix
#> Finished calculating effect matrix!